Attempt at writing the "possible variation" of #9200.
This is running aground because type synonyms need to be treated
differently than other declarations: type synonyms may result in
an *unlifted* kind, and so getInitialKinds doesn't work for them.
A little plumbing would probably fix this, but I can't sink more
time in right now. And, this problem goes away with NoSubKinds,
which is in a different branch. So, tabling for now.