Now I see why I aborted tracking boxity of Coercible.
This commit tries to change the Class stored in a ClassPred to
a tycon, in order to admit ~R#. But this is pure madness when
followed through. It might actually be saner to make a dummy
class for ~R# so that a Class can still be stored in ClassPred.

Even better: make a new constructor for PredType that classifies
representational equalities.