Abortive attempt at meta-covars.
Aborting because it's unclear what to do when we wish to assign
a covar to an RHS that isn't a Coercion (because it's an EvTerm).

There are various ways forward (just don't bind these, for example),
but I'm going to try a different tack: I believe the problem is
during zonking a *term*, when we can detect let-bound covars and
add them to a substitution to apply. This seems much simpler!