Do proper depth checking in the flattener to avoid looping.
This implements the plan put forward in comment:14:ticket:7788,
fixing #7788, #8550, #9554, #10139, and addressing concerns raised
in #10079. There are some regressions w.r.t. GHC 7.8, but only
with pathological type families (like F a = F a).

With depth checking, this commit also moves newtype-unwrapping
from the canonicalizer to the flattener. The flattener is the
right home for this logic, in my opinion.

This commit also introduces a new IntWithInf type in BasicTypes
that represents an integer+infinity. This type is used in a few
places throughout the code.

Tests in
  indexed-types/should_fail/T7788
  indexed-types/should_fail/T8550
  indexed-types/should_fail/T9554
  indexed-types/should_compile/T10139