Cache coercionKind, fixing #11598.
This commit creates a new datatype, CoercionRep, that
stores the details of a coercion. The Coercion datatype now
stores its kind, role, and free variables, so looking up
these details is very quick. In the cases that we never need
these details, I believe that laziness will prevent us from
calculating them.

This includes an update for the Haddock submodule.