Add more levity-polymorphism checks
This fixes #12973 and implements the two checks described
in the Levity Polymorphism paper: no levity-polymorphic (LP)
binders and no levity-polymorphic arguments.

See Note [Levity polymorphism checking] in TysPrim.

* Core Lint has been updated to check for LP applications.

* LP applications are checked in the desugarer. I figured out
where to add the checks by looking for all uses of the App
constructor and working backwards.

* Previously, the check for LP binders ignored binders invented
in the desugarer. This loophole is now closed.

* Class members may no longer be LP.

There are two remaining problems:

* Arrows are fully broken. The problem is that desugaring arrows
uses fixM (in dsfixXXX) and thus examining the type of a desugared
expression causes us to hang. I don't know how to proceed here.

* Error messages are less-than-ideal. But I'm not expecting
LP errors to come up often. As usual, it would take simply more
plumbing to improve these, but I'm a bit short of lead pipe at
the moment.